Make IP label clearer
From a noob standpoint, there's no documentation on how to set the config node, but there are inferences to a DVR-type server, as if the camera is a part of a system that has a central server where recordings are stored, which makes a noob question what should go in the IP field, especially if they've made a config mistake somewhere that prevents a successful connection to the camera.  So I thought changing "Server IP" to "Camera IP" might be just enough to assure a noob like me that they are correctly configuring the config node.

If this is inappropriate in other contexts, then a section in the readme explaining how to set up the config node might be helpful.